+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - flawfinder analyzer.
+- scan-build (clang) analyzer.
 
 ## [0.1] - 2017-06-12
 ### Added